requirements.txt
pytest
flake8
twisted
service-identity
attrs
pyyaml
msgpack-python
sqlalchemy
attrs-sqlalchemy
passlib
klein
jinja2
wtforms
multidict
markdown
mindspace-protocol
psutil
humanize
emote-utils
